Fixed event saving Mongo issue
- Trivial solution - stopped using new Event and just saved the incoming
event data in upsert (may need to do an additional check to ensure
conformation to EventSchema (can add a prilminary function to create a
new Event out of the incoming data and ensure no errors are thrown
before using the data for upsertion for example.

- @masterford be advised I had to make the listAll function in the
Eventcontroller public as the routes.config functions were calling list
(which has now been deleted I see in favour of listAll). If you want to
keep listAll private let me know and we'll find a compromise.